--- pages/about/me.js
@@ -1,28 +1,0 @@
-export default function AboutMe() {
-    return (
-        <div className="w-4/5 mx-auto">
-            <p className="text-center">Hi! My name is Anna and other than coding, here's what goes on inside my head!</p> 
-            <div className="ml-2">
-                <h1>Some Things I Care About</h1>
-                <h1>Free Palestine / End This War!</h1>
-                <h1>Mental Health</h1>
-                <h1>LGBTQ+ Rights</h1>
-                <h1>Black Lives Matter</h1>
-                <h1>Environment</h1>
-            </div>
-            <div>
-                <h1>Some Interests I have</h1>
-                <h1>Movies</h1>
-                <h1>Books</h1>
-                <h1>Podcasts</h1>
-                <h1>Music</h1>
-                <h1>Creating Art</h1>
-                <h1>Writing/Journalism</h1>
-                <h1>Theater</h1>
-                <h1>Chinese Food</h1>
-                <h1>Tarot</h1>
-                <h1>Outdoors</h1>
-            </div>
-        </div>
-    )
-}
