--- pages/about/my-journey.js
@@ -1,20 +1,0 @@
-export default function MyJourney() {
-    return (
-        <div className="w-4/5 mx-auto">
-            <h1 className="text-center">My Web Development Journey</h1>
-            <p className=" p-2 m-2 bg-lightgray">
-            On a good day, coding can be a fun little puzzle. On a bad day, coding can make you want to commit violence against the most expensive thing you own. As someone who didn’t really code before college, I found myself often fighting an uphill battle.
-
-            That's why, before my senior year, I took a gap semester. While I did have a job and some extracurriculars, without classes I was able to carve out the space to finally fill the conceptual holes I had, practice without much stress, and explore. I'm proud to share that where I once felt apprehension at the thought of opening VSCode, I now engage with it spontaneously and many times even eagerly. Below, you'll find a snapshot of my journey during this period, as well as my aspirations and plans for moving forward.
-            </p>
-
-            <h1 className="text-center">Guides I used</h1>
-            <p>Before my gap semester I thought about enrolling in a bootcamp. Luckily, I stumbled upon this reddit post that convinced me otherwise.</p>
-            <p>make checkboxes</p>
-
-            <p>I read Fundamentals of Web Development by Randy Connolly and Ricardo Hoar</p>
-            <h1 className="center">Plans Moving Foward</h1>
-            <p></p>
-        </div>
-    )
-}
